postgres: Add replication support for tsvector
This CL "allows" snapshot/replication to occur for postgres `tsvector`
types, but we intentionally throw away the data (and store NULL). This
is because Readyset has no intention of supporting `tsvector` data
types as they will never be a good candidate for caching, at least not
for a very long time.

Discarding the data allows us to import the rest of the table's data,
which can be used for proper cached queries in Readyset.

Fixes: REA-5734
Release-Note-Core: Support replication of postgres `tsvector` data
  types. Readyset interntionally discards the data as we will not
  support tsvector queries in the product.
Change-Id: I28f209a96939f1cc58fd616b5cd5e07d8c0ddb23
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9511
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>