logictests: Support postgres `tsvector` type
This patch allows the logictest to accept an incoming postgres data
type of `tsvector`. We also implement the tokio_postgres
`FromSql::from_sql_null()`, which allows us to catch a null response
value. It is assumed, as will be seen in other CLs in this chain,
that Readyset will emit a SQL `NULL` for `tsvector` data types.

Change-Id: I702b026a4cd990fcac1c5de16b1eb180faa368c3
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9510
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>