treewide: Add some cargo-udeps ignore configs
We could now conceivably use `cargo-udeps` in CI, but it does produce
false positives and is probably not of significant value outside
periodic checks.

One ignore is for a crate that we depend on only to modify its feature
flags as a convenience. The other is only used in doctests, which
cargo-udeps does not understand.

Change-Id: Ifc8d4ec0ac701fc7459c321bd59ee280a2c2e788
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9548
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI