vertical: Don't test against MySQL 5.7
MySQL 5.7 doesn't have the utf8mb4_0900_bin collation which we depend on
for this test. Note that this would work except that we don't handle
trailing spaces correctly; see REA-5754. For 5.7, all collations (except
maybe `binary`) are `PAD SPACE`. We can run this against 5.7 again
if/when we fix REA-5754.

Note that we already don't run this against 5.7 in nightly anyway, but I
had changed the collation to allow it to run against 5.7 as part of
testing 50ee4c8543ec3b3ec95429cdb8036171d9e85ca8. Little did I know
(because I had attempted to forget the existence of the collation
padding attribute), this collation was not the same, so the test broke.

Change-Id: I2ce6ef6d324f406a8e93987bf9fc3cfe30c15e7a
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9554
Tested-by: Buildkite CI
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>