buildkite: Add multi-arch base image build steps
Build both x86 and arm64 versions of the ubuntu22-based
"builder-base-image".  This involves creating a new pipeline with
separate steps to build the image for each architecture, along with a
single step to push the manifest that combines them.  This one pipeline
can be included in both the integration and nightly builds to avoid
duplication.

Change-Id: I26c2443527fb1be8abc491f79bdf9c2e62c83896
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9585
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI