mir: fix topk handling in `add_bogokey` rewrite
Because TopK was handled in the same function as the leaf
nodes, not all TopK cases were correctly handled.

This change splits the logic over two functions to ensure
correct handling of TopK nodes. Also introduces some refactoring
by extracting common logic into a new fn and moves the entry
point to the top of the file.

Change-Id: I54abbf32006ac93db30fc01761ef51f161921dce
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9591
Tested-by: Buildkite CI
Reviewed-by: Vassili Zarouba <vassili@readyset.io>