controller: Create a state machine of table statuses
This patch adds to the controller a simple state machine for keeping
track of table statuses.

On startup, we populate the state machine with the tables known to the
existing dataflow state, giving them an initial status of
TableStatus::Initializing or TableStatus::NotReplicated.  No other state
transitions are implemented yet, and the state machine is not yet
consumed by anything.

To update the state machine, feed a new table status into a channel.
The idea is to allow for easy sharing of the transmission side of this
channel to various parts of the system.

Each domain runs on its own single-threaded tokio runtime and is
composed almost entirely of synchronous code.  The channel comes in
handy to exfiltrate table state from places such as these, which have no
async function coloring.

All updates sent through this channel will be processed by a tokio task,
which will apply them to the state machine in the leading controller.

There is a subtlety about event ordering that is worth mentioning.

When running in standalone mode as a single process, the table statuses
are applied from the channel directly to the state machine in the order
in which they were sent.

When running in distributed mode, table statuses received through the
channel are instead sent to the leading controller's state machine via
the controller's HTTP endpoint, which does not strictly guarantee that
the updates will be processed in order, due to the asynchronous dispatch
of HTTP requests in the controller.

This doesn't really matter, since all practical deployments run in
standalone, single-process mode.  Ensuring strict ordering when running
in distributed mode is a task left for later.

Addresses: REA-2463, REA-2907, REA-4863, REA-5751
Change-Id: Ic1f653ebf1a22fdd240b34a8eacdfec3d3ffa3d1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10196
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>