Implement non-blocking SHOW READYSET TABLES with more statuses
This patch switches away from the old SHOW READYSET TABLES behavior,
which would run a blocking query on all domains for their tables'
statuses, to the new state-machine-backed SHOW READYSET TABLES behavior,
which responds immediately with the statuses of all tables by reporting
what's presently in the state machine.

The old way could block for several hours when compacting many large
tables, which was unfortunate if you were interested in seeing what the
current status was.  The blocking could also happen after creating a
cache, as the cache creation might create an index on our copy of the
table, which again holds up our single-threaded domains for potentially
long periods of time.

Previously, the reported statuses were only:

- "Snapshotting"
- "Snapshotted"
- "Not Replicated" (with reason in the description column)

Now, the reported statuses are:

- "Initializing"
- "Snapshotting" (with percentage in description column)
- "Compacting" (with percentage in description column)
- "Creating index" (with percentage in description column)
- "Online"
- "Not replicated" (with reason in the description column)

Instead of having to read logs to determine the progress of these
operations, you can now find the progress reported as percentages
displayed in the description column for the table.

Fixes: REA-2463, REA-2907, REA-4863, REA-5751
Release-Note-Core: Overhaul SHOW READYSET TABLES to always respond
  immediately, report more table statuses such as compaction and index
  creation, and report progress as a percentage in the description
  column during snapshot, compaction, and index creation.
Change-Id: I094cf7e0e9aa75fb65ea036abd94db2ed5240a08
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10207
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>