readyset-server: add support for ALTER TABLE ADD KEY
This commit adds support for the ALTER TABLE ADD KEY statement to
not require a re-snapshot of the table.
When we restart readyset we essentially re-run the snapshot for all
tables. Then we check if the body of the table we have stored in the
registry is the same as the body from the DDL we are trying to apply.
If it is, we don't need to re-snapshot the table.
The change on this commit ensures that we update the in-memory copy of
the DDL we have stored and also persists a new copy of the table body
with the new key in the registry, to avoid re-snapshotting the table
when we restart readyset.

Fixes: REA-5753
Fixes: #1521

Release-Note-Core: Avoid re-snapshotting tables when user executes
  an ALTER TABLE ADD KEY statement.

Change-Id: If42bc9b9ba78179fe22e7f6cb8bb030f90e1e713
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9601
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI