buildkite: Create multi-arch readyset-build containers
Create readyset-build containers for both x86 and arm for use in test
steps:
- update Dockerfile to support both architectures (as well as some minor
  improvements that should help with layer caching).
- create a new `pipeline.build-image.yml` pipeline to avoid having to
  duplicate the multi-arch image build and manifest create/push steps.
- replace old single-arch image build steps with ones that upload the
  new pipeline.

Unfortunately, sub-steps in the monorepo-diff pipeline which have
dependencies on the build-image pipeline upload step don't get scheduled
properly by buildkite.  So the monorepo-diff block had to be split out
into two (one with build-image dependent steps and one without) so that
the build-image dependency could be added to the entire block.  (See
comments in pipeline.core.yml for more information.)

Change-Id: I6a6a6964853b91079b9047427ec24d362e48dabb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9618
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>