authority: Improve leader tracking in standalone
When initializing the leader state in the standalone authority, we now
initialize last_leader to ourselves if we won the election and there was
no prior leader.

Two AuthorityUpdate message types are worth mentioning,
WonLeaderElection and LeaderChange.  WonLeaderElection tells the
controller that it is now the leader.  LeaderChange tells the controller
that some other controller won the election.

Previously, this state tracking would tell the controller first that it
won the election via WonLeaderElection and then that some other
controller was elected via LeaderChange, even though that controller is
actually ourself.

This wasn't really a problem, because we'd send the same controller URL
in the LeaderChange payload.

Addresses: REA-2463, REA-2907, REA-4863, REA-5751
Change-Id: Ibbedeb190205c071968ad36f8f14d26fd1a69c4f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/10194
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI