readyset-data: Potentially coerce type for all collations
We adjusted for citext only, but all collations are worthy of coercion.

I suspect that this is not really the right place for this code and that
we need to have a lecture on where type promotion should be, but I also
don't think today is that day.

Note that tests that circumvent the frontend will need to be more
careful in the future to pass in the correct collation when passing in
text values.  Many sloppy things that worked before will no longer be
kosher.

Change-Id: I65da3d6a565332a77138b5711adb9304fef5c52f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9620
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>