readyset-adapter: Add controller-side filtering to verbose_views
Previously, we would send the entire list of caches to the adapter,
which then might perform additional filtering.

Now, we add 2 optional parameters to verbose_views for filtering by
query_id and name.  If both parameters are empty, we return the full
list.  If set, we return the caches that match at least one of them.

Addresses: REA-5765
Change-Id: I813729447dc34e64bad4555290565f62609770eb
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9654
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI