readyset-adapter: Improve duplicate cache creation handling
Previously, before creating a named cache, we would first drop any
existing cache with that same name before creation.  In Iad22f58, we
added a similar check for unnamed caches, comparing query_id to first
drop a duplicate before creation.

It's important to first drop a duplicate cache to prevent the graph
metadata in SqlToMirConverter from erroneously being updated with node
index values that are then never persisted to the graph itself, which
can happen if we decide to abandon the in-progress graph creation to
instead reuse the previous cache.

Unfortunately, we could still run into this inconsistency between
SqlToMirConverter and the graph when creating duplicate caches with
different names.

Now, instead of these targeted checks that made decisions based on what
was in the adapter's view name cache, we instead query the controller
for any existing caches that match either the same query_id or requested
name.  If any such cache exists, drop it before creation.

Fixes: REA-5765
Release-Note-Core: Fix a bug that led to errors when attempting to
  create duplicate caches.
Change-Id: I4e5aa25d19e1ea4488c986ee65f4a85ec8a801a5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9655
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>