readyset-sql, replicators: Fix build errors
We made two changes in CLs that were tested and merged concurrently with
a new [CL 9601] that had compilation errors when rebased on those
changes (described below) . Because the codebase won't compile without
fixes for both issues, this patch combines both to get Buildkite
verification to succeed and thus allow Gerrit to merge the CL.

[CL 9583] removed imports of `test_utils::slow` in favor of the new test
tag, while CL 9601 introduced two tests using the old style `#[slow]`
(likely to be removed entirely in the future).

[CL 9541] removed `constraint_name` from `TableKey::Key`, which caused
the introduction of `set_constraint_name` in CL 9601 to be out of sync.

Note that there is no `CONSTRAINT <constraint_name> KEY <index_name>`
syntax in either MySQL or PostgreSQL, since non-unique keys do not imply
a constraint.

Additionally, there was an incorrect comment stating that
`TableKey::CheckConstraint` does not have a constraint name in MySQL, so
I removed that. I'm not sure our parsing of it in nom-sql is correct,
but clearly if not it isn't tested or relied on by customers. But
sqlparser parses it correctly and we already implemented converting it.

[CL 9601]: https://gerrit.readyset.name/c/readyset/+/9601
[CL 9583]: https://gerrit.readyset.name/c/readyset/+/9583
[CL 9541]: https://gerrit.readyset.name/c/readyset/+/9541

Change-Id: I6a6a6964f36f05bee75e5f683e62282f0000283c
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9648
Reviewed-by: Marcelo Altmann <marcelo@readyset.io>
Tested-by: Buildkite CI