readyset-sql-passes: Parse some rewritten Postgres queries as MySQL
This allows us to switch these tests to using `readyset-sql-parsing`
instead of `nom-sql`, so we also remove `nom-sql` from the dependencies
for `readyset-sql-passes`.

We rewrite `WHERE y IN ($1, $2, $3)` to `WHERE y = ?`, and verifying
this rewrite in tests attempts to parse the expected query string. Doing
so with Postgres fails because Postgres does not support the `?`
placeholder, but our MySQL parsing and conversion from sqlparser is
flexible enough to allow for both dollar sign and question mark
placeholders in the same query.

(Historical note: sqlparser probably cannot support both dollar sign and
question mark placeholders in the same query *for Postgres*, because
Postgres has a `?` binary operator, which requires changing both the
precedence and tokenization rules that affect parsing `?`. The generic
and MySQL dialects are able to handle both because there is no `?`
binary operator.)

Addresses: REA-5049

Change-Id: I6a6a6964d88147d61ed8232eabc96f46d5115157
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9669
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI