replicators: Don't use serde for parsing Postgres DDL events
The previous approach was to hack in a SQL parsing pass as part of
deserialization of certain fields in our custom JSON messages for
replicated DDL events. This was a bit cumbersome and magical, for no
real benefit. Additionally, it prevents paramaterizing the parsing with
`ParsingMode`.

Now, we defer parsing the fields that contain SQL until converting the
DDL event into a `Change`, and make that operation fallible.

Change-Id: I6a6a696422e0be1f1659a8238e424c2421436cdd
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9682
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI