topk: fix wrong results on non-unique order col
Every now and then, we would get incorrect results
when the query contains an order by on a non-unique column.

After dropping to m values (assuming m < k), we lookup the group
by key in the parent, sort the result set, skip the first m and take
k - m. That sounds reasonable when the order by is unique.

However, in the case of non-unique order by cols, the sort on the
result set doesn't guarantee that the top m records (which we skip)
match what we currently have in the node. This can cause
duplicates and ultimately incorrect state.

Change-Id: I119e0648ac6722bf57146af545e4172d9a72b7f5
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9689
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>