readyset-sql: Convert sqlparser representation of prefix key parts
We attempt to throw away everything except the column name. Whether it
is worth preserving the prefix expression or generic expressions
("functional key parts") is a question for the future. nom-sql currently
doesn't support it, so we aren't losing anything in the switch to
sqlparser. However, some tables won't snapshot if they have functional
key parts because, unlike `ALTER`s, we don't have a partial/fallible
representation for `CREATE TABLE`. So we may want to support that.

Note that this involves updating sqlparser to a version that supports
prefix key parts.

Fixes: REA-5769

Change-Id: I6a6a6964b590098139237921924c7fc6dd6c1125
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9694
Tested-by: Buildkite CI
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>