readyset-sql-parsing: Rate limit mismatch logging
We apply this rate limiting to places where we might be parsing selects
(e.g. ad hoc query handling and prepare), but if we make it to
`/extend_recipe` or encounter something we parse differently in
replication, we don't rate limit this logging.

The default is 60 seconds and can be configured via a
`PARSING_LOG_RATE_LIMIT_SECS` environment variable. Unlike the parsing
config (which needs to be different or modified at different callsites),
this rate limit duration configuration is not worth plumbing through as
a CLI flag.

Fixes: REA-5047
Release-Note-Core: Rate-limit logging of nom-sql/sqlparser-rs
  mismatches, configurable by the `PARSING_LOG_RATE_LIMIT_SECS`
  environment variable. By default, we will log a warning for mismatches
  at most once per minute. If the environment variable is set to `0`, no
  rate limiting will be applied. The rate limit does not apply to DDL
  encountered during replication.

Change-Id: I6a6a69642fce2265cbaf45f347f44bdf72b59ed1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9702
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>
Tested-by: Buildkite CI