Straddled Joins: Add flag for full mat right side
Add the necessary plumbing to pass a flag from mir to join node
indicating if the right side is fully materialized.
When the right side is partial, we might miss on lookup.

At the moment, we always assume the right side is not fully
materialized. This will be changed once we implement the new sj
algorithm.

Change-Id: I9ad34a728386cc37467bb82f3b0e9b3a8fb53ded
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9711
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>