readyset-sql-parsing: Only log certain mismatches in adapter
We introduce yet another field on the config to allow the adapter to be
more selective about which types of statements will log on a
sqlparser-rs/nom-sql mismatch. Specifically, it allows for only logging
mismatches for `SELECT` statements and Readyset extensions (importantly,
`CREATE CACHE`). So, for example, we don't care about supporting the
`RETURNING` clause in `INSERT` statements, so if sqlparser successfully
parsed that, and nom didn't, we would ignore the mismatch (as far as
logging: if the config is set to panic, we will still panic).

Change-Id: I6a6a6964a109483236dccc23cc31f07a2f2db75b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9704
Tested-by: Buildkite CI
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>