replicators: Save replication offset for freshly-created MySQL tables
We were not consistently saving the offset due to the potential for an
empty schema field in Change::CreateTable, which would then cause the
lookup into the table's mutator to fail.

Set the schema field if it was empty.  Also consider tables added so
long as they are not immediately dropped in the same batch of events.

Fixes: REA-5787
Release-Note-Core: Fix a bug where Readyset would fail to save the
  replication offset for MySQL tables created during active replication,
  leading to unnecessary table resnapshots.
Change-Id: I45665deb0d85629e331fc9066eab861eee5127f2
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9723
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI