readyset-server: Add magic number to inter-domain communication.
Currently, the inter domain communication and direct table
communication used by replicators is done using a tcp connection
socket. It's possible for someone to connect on one of the ports and
send arbitrary data to the server. This causes the whole domain to
fail and get replaced.

This commit adds a magic number to the inter-domain communication
and direct table communication. This magic number is used to
identify the connection as originating from Readyset.

Release-Note-Core: Add magic number to inter-domain communication to
  prevent arbitrary connections from being accepted.

Fixes: REA-5788
Change-Id: Idb7aba257c0df5349fbe6f7c5585e2536cf60109
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9734
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI