replicators: When restarting replicator, log the position we were at.
When we restart the replicator, we start from the last position we
applied to a table. In case of an instance where we do not snapshot
all tables, the position we error out vs the position we restart at
might be different. For a better debugging experience, log the
last parsed position before erroring out.

Release-Note-Core: Improve logging when restarting replicator to
  print the last replication position.
Change-Id: I1fac045c7253e8ca141b295899618d61226369c1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9739
Tested-by: Buildkite CI
Reviewed-by: Michael Zink <michael.z@readyset.io>