antithesis: Add images and config for QueryPilot
This is currently configured to just run SQLancer against ProxySQL and
run QueryPilot at a fixed interval. It runs in a separate container from
ProxySQL, which may not be how we usually deploy it, but it avoids
building a proxysql+query-pilot image, and will make scheduling/running
QueryPilot more flexible on the Antithesis platform (e.g. when doing
multiverse debugging).

Change-Id: I6a6a6964fe5514cdeb7b822dd70e5658f93e2d42
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9768
Reviewed-by: Sandaruwan Gunathilake <sandaruwan@readyset.io>
Tested-by: Buildkite CI