readyset-sql: Ignore unparsed part of `EXPLAIN CREATE CACHE` in equality
This is already what we do for real `CREATE CACHE` statements, and this
avoids this hack of stripping off the `EXPLAIN` keyword to match
nom-sql. That hack was problematic because it couldn't handle non-space
whitespace, e.g. `"EXPLAIN\nCREATE CACHE"` should work but would instead
panic.

We also rework the parsing to use the same top-level parsing routine as
for real `CREATE CACHE` statements, which allows us to support its
options (`ALWAYS`, `CONCURRENTLY`, and cache name) when parsing the
`EXPLAIN` version as well. This way you can always just put `EXPLAIN` in
front of your `CREATE CACHE` statement. nom-sql doesn't support this.

Fixes: REA-5821

Change-Id: I6a6a6964b2df4061fc945273829ddedfba4ea522
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9772
Tested-by: Buildkite CI
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>