readyset-client: Report queries as supported after cache drop
This commit is basically a revert of Ia1a6bc83.

Before this revert, dropping a cache would render its query as
unsupported via EXPLAIN CREATE CACHE, even though the query is actually
supported.

The idea in the original commit was to provide a mechanism to add
queries to a denylist when running with --query-caching in-request-path.
In this mode, Readyset automatically caches queries it sees, and the
intent was to provide a mechanism to prevent selected queries from being
cached.

The idea may have been nice, but this behavior was never documented,
only takes effect when DROP CACHE specifies a cache by name, and does
not take effect when DROP ALL CACHES is executed.

Perhaps most importantly, Readyset is not run in this implicit caching
mode for real workloads.  Let's revert this hidden feature to make our
reporting of query support via EXPLAIN CREATE CACHE actually correct.

Fixes: CLD-1309
Change-Id: Ib113b17140a5c392c5da7f9dde6286e60f657b07
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9773
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI