dataflow: Refactor Join on_input
Refactor the join on_input execution code to make it easier to reason
about.

Changed some variable names, adopting the terminology of This Side
(ts) to refer from columns/keys coming from the side of the replay and
Other Side (os) to for the side of the join that is not the replay.
Also, keys is being used interchangeably with cols. Now cols refers
to the column position we indexed on and keys refer to the actual
values populated for the columns.

We will have 3 types of execution modes:
- RegularLookup: When we have predicates only in one side of the join.
  Once the replay arrives, we lookup the ON keys in the rhs and
  generate the result set.
- StraddledHashJoin: Current mode for straddled joins. We trigger two
  upquery, one for each side. When the first upquery replay arrives,
  we buffer the result. Once the second upquery replay arrives, we
  perform a hash join on both resultsets. This mode will continue to
  be used for joins where the rhs is not fully materialized.
- StraddledLookup (not implemented yet): When the rhs is fully
  materialized, we trigger an upquery in the lhs and save the
  predicates from the rhs. Once the replay arrives, we lookup the
  predicates in the rhs and build a new set of predicates combining
  the ON keys + the original predicates.
  Then we execute a state lookup like RegularLookup does.

Change-Id: I6d80ed76c520e0467a16c2729bfbd0a4c74a7ba1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9777
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>