dataflow: Optimize Straddled Join Performance.
For joins where the rhs is fully materialized, optimize the join
execution by triggering a single upquery on the lhs. This is done when
an upquery misses. In that case, we get the predicates from each side
of the join, we save the predicates from the rhs in a temporary buffer
and execute the upquery on the lhs only. Once the replay for the lhs
upquery arrives, we do a lookup based on the predicates from the rhs
saved in the temporary buffer.

When all predicates are equality, we use Index Condition Pushdown (ICP)
to push both ON and Predicates down to RocksDB in order to do the rhs
lookup. For range queries, we do a lookup based on the ON conditions
and resolve the range predicates in memory.

In case of rhs being partial, we fallback to old join execution of
triggering upqueries on both sides with just the predicates and later
doing a hash join to resolve the join.

The new join algorithm is more efficient specially when the rhs has
low cardinality.

On test using customer data, the performance improved from 1QPS (old
algo) to 2907QPS (new algo), a 2907x improvement.

Release-Note-Core: Introduce a new join algorithm for straddled joins
  (joins where the predicates come from both sides of the join). The
  new algorithm uses optimizations like Index Condition Pushdown (ICP)
  to push down the predicates to RocksDB, resulting in faster lookups
  and reduced memory footprint.

Fixes: REA-5027
Change-Id: I3e4028fcb0e799edd08f0424ef84372d28ceb96f
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9713
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI