buildkite: Fix problem pushing latest docker images
The logic was backwards in a test that meant when building multi-arch
images, the `latest` tag was being pushed on the image itself.  Later,
the multi-arch manifest gets pushed at that tag.  But it's possible to
time it such that the single-arch image at `latest` gets pulled before
the multi-arch manifest at `latest` is pushed.  This was causing weird,
occasional failures.

Change-Id: I6a6a696444bd00d9bf30ffb5b2ba20f0563c5b4e
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9808
Tested-by: Buildkite CI
Reviewed-by: Gautam Gopinadhan <gautam@readyset.io>