readyset-sql: Represent numeric/decimal literals as strings
I don't know the rationale for the previous `(mantissa, scale)`
representation, but it doesn't have anything to do with the syntax that
was input, necessitates a redundant re-parse step, and makes it
impossible to parse or generate some larger numerics which we now
support (i.e. if the mantissa doesn't fit in an `i128`—basically the
same limitation we had previously).

Change-Id: I6a6a69647a00fac45a3d174f10ff30c3e58b4e99
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9824
Tested-by: Buildkite CI
Reviewed-by: Mohamed Yasser Abdelhamed Abdeen <mohamed@readyset.io>