build: (rpm) Avoid overwriting modified conf files
By default, rpm will rename an existing conf file and install the one in
the RPM.  It is better for users if their existing conf file is left
alone.  The new one will be installed, but as <file>.rpmnew.

This update also adds some helpful references to the
`cargo-generate-rpm` and `cargo-deb` github projects.

BuildDistroPkg: donotuse
Change-Id: I6a6a6964297cb71ef3bba9f0011f2c2fd88a5df1
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9843
Tested-by: Buildkite CI
Reviewed-by: Gautam Gopinadhan <gautam@readyset.io>