treewide: Update chrono-tz
A [recent update to the Iran timezone][0] in the IANA Time Zone Database
moved the half-hour shift in Iranian time from starting 1971-01-01 to
starting 1978-11-11, causing us to not match times at the end of 1978 if
the upstream's TZ database was more up to date (i.e. 2025b). As we don't
currently have a way to match the TZ database used by the upstream, we
may as well be as up to date as we can.

However, 2025b also [obsoletes MST, CET, etc.][1] which "were present
only for backward compatibility with UNIX System V, last released in the
1990s." They are now aliases for 'real' timezones (e.g. `EST` maps to
`America/Panama` and `EST5EDT` to `America/New_York`), which differ from
the previous legacy timezones when you go back a few thousand years.
This would be fine, except that there is still a mismatch between
`chrono-tz` and Postgres despite both using version 2025b of the TZ
database; Postgres seems to be forcing the old behavior or resolving the
alias to the old timezone definition. (`chrono-tz` [seems][2] to
correctly use the new alias.) So since they are obsolete anyway, we
remove them from this proptest.

[0]: https://github.com/eggert/tz/commit/4878fa40ec93499d3a7e265cd8e0ff52237ada75
[1]: https://github.com/eggert/tz/commit/a0b09c0230089252acf2eb0f1ba922e99f7f4a03
[2]: https://github.com/chronotope/chrono-tz/commit/711baee12c15dae484d69a3f75349a276b097dde

Addresses: REA-4425

Change-Id: I6a6a6964e8947acc8590a0c9b79acac3b2878eab
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9886
Reviewed-by: Jason Brown <jason.b@readyset.io>
Tested-by: Buildkite CI