readyset-sql, nom-sql: Avoid overflow when negating numeric literal
We have an (arguably ill-placed) negation-folding optimization in
nom-sql, which we replicate in readyset-sql when converting
sqlparser-rs's AST. The current implementation can attempt to negate an
unsigned integer that is too large to fit in a signed integer. In this
case, we can still perform the folding by turning it into a negative
numeric literal.

The alternative would be to not do the folding (leaving it as a negation
of a literal), but that would make the AST produced dependent on the
values of literals which would probably have unexpected results; it also
might unnecessarily limit the literals we are able to autoparameterize,
and make autoparameterization dependent on the specific value of the
literal, which also seems ill-advised.

FYI this was found using generative testing of ASTs.

Change-Id: I6a6a6964824bf23ff02fdade5e26c5e5d03da7a0
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9891
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI