readyset-sql, nom-sql: Fix parsing of function calls without parens
Some built-in functions are called without parens and cannot be used
as identifiers without escaping.

- We were missing a few (such as Postgres-specific `USER`)
- We had misclassified `NOW`, which cannot be called without parens in
  either dialect.
- We were incorrectly parsing them and then failing if they were at
  the beginning of otherwise legitimate identifiers; i.e. `WHERE
  username = 'foo'` would fail when encountering `name` instead of
  whitespace or EOF.
- The AST could not represent this difference, so `CURRENT_TIMESTAMP`
  would display as `CURRENT_TIMESTAMP()`.

In this case, it was easiest to fix the nom-sql parsing. However, fixing
the AST representation was trickier. Basically, we make f unction
arguments an `Option<Vec<Expr>>` instead of `Vec<Expr>`. sqlparser-rs
does something similar, but even more explicit (with an enum for
`FunctionArguments::List`, `FunctionArguments::None`, and even
`FunctionArguments::Subquery`). This steps on a lot of code which
expects to be able to iterate over the arguments (e.g. visitors), and
made a lot of code a lot more verbose. It's probably the most correct
though.

Change-Id: I6a6a6964b05ac5059de1f1d210e1567dff5f4ded
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9893
Tested-by: Buildkite CI
Reviewed-by: Sidney Cammeresi <sac@readyset.io>