readyset-sql: Remove `Literal::Float` and `Literal::Double`
It is extremely ill-advised to infer type information during parsing.
The AST should really represent numbers as strings, and those can be
converted as needed (and sometimes we can do so with more type
information).

We take one small step in the right direction by removing
`Literal::Float` and `Literal::Double` and leaving numeric literals as
`Literal::Number` (renamed from `Literal::Numeric` to reduce confusion
with known-`NUMERIC`/`DECIMAL` types, and chosen to be similar to the
sqlparser-rs AST type). Then, we use the dialect to decide whether it is
a `DOUBLE` (MySQL) or `NUMERIC` (PostgreSQL) once we need to: e.g. when
lowering expressions in dataflow and when constructing a `ViewQuery`.
This, for better or worse, requires plumbing the dialect through to
several places it didn't exist before.

Change-Id: I6a6a6964fb02c03c31c7c4482ce370f540be2483
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9892
Reviewed-by: Sidney Cammeresi <sac@readyset.io>
Tested-by: Buildkite CI