benchmarks: fix polling for readyset ready
At benchmark start time, we check to see if readyset is done with
snapshotting the database. We do this by calling `SHOW READYSET
STATUS`, and looking at the value of the `Status` row (where `Status`
is the value in in 0th column). To find the `Status` row, we were
coercing `b"Status"` to a vec, and relying on the old DfValue coercion
to try a byte array into a UTF-8 string. That coercion was removed,
and thus is benchmark couldn't find the status.

This patch simply removes the "string to byte array", and just coerces
a string to a DfValue.

Change-Id: Ie8e66b80f55dab26636d1048417ea0adafaf1be9
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9934
Reviewed-by: Johnathan Davis <jcd@readyset.io>
Tested-by: Buildkite CI