controller: DfType for custom types' default value
If a postgres table's column has a custom type (which includes enums),
Readyset does not resolve that custom type when determining the
column's default value at lowering from MIR to Dataflow. This causes a
failure when deriving the column's default value when the column is
attributed as either having a default or being not null.

This CL passes the known `custom_types` to the existing call to
`DfType::from_sql_type()`.

Addresses: REA-5746
Release-Note-Core: Fix that allows snapshotting postgres tables where
  a column of a custom type is attributed NOT NULL but has no default
  value.
Change-Id: I4e37e44bbe0b56b120ccfdb7d65e702e5d1afd8b
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9954
Reviewed-by: Michael Zink <michael.z@readyset.io>
Tested-by: Buildkite CI