readyset-sql, nom-sql: Support `CONVERT` expressions
In order to maximize support while minimizing effort, we represent
`CONVERT(expr, type)` expressions with `Expr::Cast` so they will be
supported automatically in dataflow. We do so by replacing the
`postgres_style` flag with a `CastStyle` enum.

`CONVERT(expr USING charset)` is represented with a new
`Expr::ConvertUsing` variant, but will require future work to support in
dataflow.

Change-Id: I6a6a6964a59b124bc41e0e618868e0b529f7ad4d
Reviewed-on: https://gerrit.readyset.name/c/readyset/+/9993
Tested-by: Buildkite CI
Reviewed-by: Jason Brown <jason.b@readyset.io>